% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AcuteLiverFailure.R
\docType{package}
\name{AcuteLiverFailure-package}
\alias{AcuteLiverFailure}
\alias{AcuteLiverFailure-package}
\title{AcuteLiverFailure: A Study Package for Generating Cohort Diagnostics}
\description{
A package that generates cohort diagnostics for cohort definitions included in the package.
}
\keyword{internal}
